Refactor PhaseColumn and update Arrow component
Consolidate verification chain components:
- Rename PhaseColumn.tsx to Phase.tsx
- Update import references in VerificationPhases
- Remove unused description property in CMS steps
- Simplify Arrow component with reusable SVG rendering
- Restrict direction type to "down" and "right"
